Menu bg depends on menu_mode and honors config
The menu bg color will now honor the setting in the config file if the
menu mode is normal. But if menu mode is confirm-delete it'll be red.
This restores the original behavior where bg color of the menu depends
on menu_mode, while still allowing the color for normal mode to be
configured by the user in the config file.